requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.7.1
alembic==1.13.1
asyncpg==0.29.0
python-multipart==0.0.20